git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@1370 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ base64.c
@@ -31,7 +31,7 @@ static void base64_init(void)
     return;
 }
 
-int base64_decode(byte *buff, size_t size)
+size_t base64_decode(byte *buff, size_t size)
 {
     static int table_set_up = 0;
     size_t count = 0;
